# springboot_openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
paths:
...
  /auth/register:
    post:
      tags:
      - auth-controller
      operationId: customerRegister
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAORegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Violazione di un vincolo nella base di dati
          content:
            application/text:
              schema: {}
        "409":
          description: Username già in uso
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: customerLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAOLoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Sintassi dell'input non valida
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /suspensions/byUser/{username}/suspended-permanently:
    get:
      tags:
      - suspension-controller
      operationId: isPermanentSuspended
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /auth/token-check:
    get:
      tags:
      - auth-controller
      operationId: tokenCheck
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /auth/suspended-login:
    get:
      tags:
      - auth-controller
      operationId: suspendedLogin
      parameters:
      - name: permanently
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /auth/manager-login:
    get:
      tags:
      - auth-controller
      operationId: managerLogin
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /auth/admin-login:
    get:
      tags:
      - auth-controller
      operationId: adminLogin
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
components:
  schemas:
...
    DAORegisterRequest:
      required:
      - cf
      - email
      - gender
      - name
      - password
      - surname
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 4
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
        name:
          maxLength: 50
          minLength: 4
          type: string
        surname:
          maxLength: 50
          minLength: 4
          type: string
        gender:
          pattern: "[MFX]"
          type: string
        cf:
          maxLength: 16
          minLength: 16
          type: string
        email:
          type: string
    DAOLoginRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
...